Move warehouse and mirrored warehouse clients outside their subfolder (no needed)
